+++ src/aarkanoid/gameObjects/MovableObject.java
@@ -3,15 +3,17 @@ package aarkanoid.gameObjects;
 import java.awt.*;
 
 public abstract class MovableObject extends GameObject {
-    protected float dx, dy;
+    protected double dx, dy;
 
-    public MovableObject(int x, int y, int width, int height, float dx, float dy) {
+    public MovableObject(int x, int y, int width, int height, double dx, double dy) {
+
         super(x, y, width, height);
         this.dx = dx;
         this.dy = dy;
     }
 
-    public MovableObject(int x, int y, int width, int height, float dx, float dy, String texturePath) {
+    public MovableObject(int x, int y, int width, int height, double dx, double dy, String texturePath) {
+
         super(x, y, width, height, texturePath);
         this.dx = dx;
         this.dy = dy;
@@ -24,14 +26,14 @@ public abstract class MovableObject extends GameObject {
         bounds.setLocation(x, y);
     }
 
-    public void setVelocity(float dx, float dy) {
+    public void setVelocity(double dx, double dy) {
         this.dx = dx;
         this.dy = dy;
     }
 
-    public float getDx() { return dx; }
-    public float getDy() { return dy; }
+    public double getDx() { return dx; }
+    public double getDy() { return dy; }
 
-    public void setDx(float dx) { this.dx = dx; }
-    public void setDy(float dy) { this.dy = dy; }
+    public void setDx(double dx) { this.dx = dx; }
+    public void setDy(double dy) { this.dy = dy; }
 }
